# Core: Upgrade docs-mdx for smaller install