--- a/code/core/package.json
+++ b/code/core/package.json
@@ -277,7 +277,7 @@
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-scroll-area": "^1.0.5",
     "@radix-ui/react-slot": "^1.0.2",
-    "@storybook/docs-mdx": "4.0.0--canary.19.48b6c75.0",
+    "@storybook/docs-mdx": "4.0.0--canary.19.f920eba.0",
     "@storybook/global": "^5.0.0",
     "@storybook/icons": "^1.2.5",
     "@tanstack/react-virtual": "^3.3.0",