Closes N/A

## What I did

Upgrade `@storybook/docs-mdx` to an ESM-only version that no longer uses babel and runs asynchronously

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [x] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

Fire up a sandbox and checks the docs/mdx stories

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-28552-sha-be84deb3`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-28552-sha-be84deb3 sandbox` or in an existing project with `npx storybook@0.0.0-pr-28552-sha-be84deb3 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-28552-sha-be84deb3`](https://npmjs.com/package/storybook/v/0.0.0-pr-28552-sha-be84deb3) |
| **Triggered by** | @shilman |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`shilman/optimize-docs-mdx`](https://github.com/storybookjs/storybook/tree/shilman/optimize-docs-mdx) |
| **Commit** | [`be84deb3`](https://github.com/storybookjs/storybook/commit/be84deb3d77d177b0db65442e6830ffd54e52b84) |
| **Datetime** | Thu Jul 11 16:28:00 UTC 2024 (`1720715280`) |
| **Workflow run** | [9895111271](https://github.com/storybookjs/storybook/actions/runs/9895111271) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=28552`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | % |
| ---- | ------ | ----- | ---- | - |
| createTime |  18.5s | 19.2s | 692ms | 3.6% |
| generateTime |  21.7s | 21.6s | -85ms | -0.4% |
| initTime |  24s | 22.9s | -1s -17ms | -4.4% |
| createSize |  0 B | 0 B | 0 B | - |
| generateSize |  76.5 MB | 76.5 MB | 104 B | 0% |
| initSize |  205 MB | 198 MB | -6.42 MB | -3.2% |
| diffSize |  128 MB | 122 MB | -6.42 MB | 🔰-5.3% |
| buildTime |  14.5s | 14s | -523ms | -3.7% |
| buildSize |  7.59 MB | 7.59 MB | 0 B | 0% |
| buildSbAddonsSize |  1.63 MB | 1.63 MB | 0 B | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | 0% |
| buildSbManagerSize |  2.3 MB | 2.3 MB | 0 B | 0% |
| buildSbPreviewSize |  349 kB | 349 kB | 0 B | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - |
| buildPrebuildSize |  4.47 MB | 4.47 MB | 0 B | 0% |
| buildPreviewSize |  3.12 MB | 3.12 MB | 0 B | 0% |
| testBuildTime |  0ms | 0ms | 0ms | - |
| testBuildSize |  0 B | 0 B | 0 B | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - |
| devPreviewResponsive |  8.4s | 8.3s | -10ms | -0.1% |
| devManagerResponsive |  6s | 5.8s | -248ms | -4.2% |
| devManagerHeaderVisible |  744ms | 830ms | 86ms | 🔺10.4% |
| devManagerIndexVisible |  775ms | 867ms | 92ms | 🔺10.6% |
| devStoryVisibleUncached |  1.4s | 1.1s | -326ms | 🔰-27.9% |
| devStoryVisible |  795ms | 901ms | 106ms | 🔺11.8% |
| devAutodocsVisible |  678ms | 852ms | 174ms | 🔺20.4% |
| devMDXVisible |  699ms | 760ms | 61ms | 🔺8% |
| buildManagerHeaderVisible |  736ms | 853ms | 117ms | 🔺13.7% |
| buildManagerIndexVisible |  739ms | 856ms | 117ms | 🔺13.7% |
| buildStoryVisible |  789ms | 906ms | 117ms | 🔺12.9% |
| buildAutodocsVisible |  644ms | 699ms | 55ms | 🔺7.9% |
| buildMDXVisible |  637ms | 754ms | 117ms | 🔺15.5% |

<!-- BENCHMARK_SECTION -->